chore: split token enum into separate file
